Fix a mis-use of curl library
But a not really obvious one that could have been documented!
curl_multi_socket_action, can close and open new connection. When it is
actually done, the socket should be given to the current HTTPP
connection. It seems to happen when a "big" request is performed and
when curl request a chunked transfer.