Improve the read_everything + fix a bug in the readBody
The readBody was returning an EOF to indicate that we finished to read
the body, however, if a client close the connection in the middle of an
upload, an EOF is also returned.
All the code has been migrated to use the read_everything() function.
The lambda is not that nice to read, but with generic lambda it is not
that bad.